airflow/requirements.txt
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.9.0
apache-airflow-providers-amazon==8.5.1
apache-airflow-providers-http==4.6.0
psycopg2-binary==2.9.9
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0
great-expectations==0.17.12
dbt-postgres==1.7.0
minio==7.2.0
